requirements-base.txt
psycopg2 >= 2.5.4
-e git://github.com/mysociety/mapit.git@master#egg=django-mapit
django-user-accounts==2.0.3
django-appconf==1.0.3
django-mailer==1.2.2
mock==1.3.0
mockredispy==2.9.0.12
redis==2.10.5
django-apptemplates==1.4
stripe==2.35.0
six

# Bulk lookup
django-formtools==2.0
defusedxml==0.5.0
pyexcel==0.4.5
pyexcel-io==0.3.4.1
pyexcel-odsr==0.3.2
pyexcel-xls==0.3.3
